requirements.txt
aiofiles==23.2.1 ; python_version >= "3.10" and python_version < "4.0"
altair==5.3.0 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.3.3 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.2.1 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.110.1 ; python_version >= "3.10" and python_version < "4.0"
ffmpy==0.3.2 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.13.3 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.50.0 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2024.3.1 ; python_version >= "3.10" and python_version < "4.0"
google-ai-generativelanguage==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.18.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.18.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.29.0 ; python_version >= "3.10" and python_version < "4.0"
google-generativeai==0.4.1 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.63.0 ; python_version >= "3.10" and python_version < "4.0"
gradio-client==0.15.0 ; python_version >= "3.10" and python_version < "4.0"
gradio==4.25.0 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.62.1 ; python_version < "4.0" and python_version >= "3.10"
grpcio==1.62.1 ; python_version < "4.0" and python_version >= "3.10"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
huggingface-hub==0.22.2 ; python_version >= "3.10" and python_version < "4.0"
idna==3.6 ; python_version >= "3.10" and python_version < "4.0"
importlib-resources==6.4.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.3 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.10" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.8.3 ; python_version >= "3.10" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
numpy==1.26.4 ; python_version >= "3.10" and python_version < "4.0"
orjson==3.10.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.3.0 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.23.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.16.3 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.6.4 ; python_version >= "3.10" and python_version < "4.0"
pydub==0.25.1 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.17.2 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
pyparsing==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.34.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rich==13.7.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
rpds-py==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
ruff==0.3.5 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
semantic-version==2.10.0 ; python_version >= "3.10" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.10" and python_version < "4.0"
tomlkit==0.12.0 ; python_version >= "3.10" and python_version < "4.0"
toolz==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.66.2 ; python_version >= "3.10" and python_version < "4.0"
typer-cli==0.12.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
typer-slim[standard]==0.12.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
typer[all]==0.12.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
typing-extensions==4.10.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.29.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
websockets==11.0.3 ; python_version >= "3.10" and python_version < "4.0"
